Fixed a bug with UNION field, added performance optimizations, updated dependencies.
- updated dependencies to clojure 1.4, removed clojure-contrib dependency,
instead linked to data.json 0.1.2

- core.clj - added handling of optional fields, fixed UNION Schema bug.
To implement optional fields you have to make a union between NULL and other type
that you want to be optional. But, if in the schema there is a null field,
then old implementation throws exception if you try to put some data in there.
For example we have a schema with [NULL, Real-type] - when trying to determine
it's (optional) type - NULL always works. So we need to make additional check if
the data are really NULL.

- Some performance improvements - added persistent! (uses transients feature from
clojure 1.4), does in-place updates.

- added :dynamic metadata to named-types because named-types are changed in the code

- added sync method for flushing the file streams where avro-writer writes.
Useful to have control and  not to lose buffered data that was not yet flushed do disk.